using git pull before push - Update JenkinsFile
https://stackoverflow.com/questions/8901973/why-git-push-is-rejected-git-pull-doesnt-help

That is your problem. When you pull, Git is telling you that your branch my_branch is up to date, not master, which is behind origin/master making a fast-forward merge impossible.


In order to push master, you need to check out master and pull. This will merge in the changes waiting on origin/master and allow you to push your own changes.

git checkout master
git pull
# resolve conflicts, if any
git push